+++ NEWS.md
@@ -1,7 +1,7 @@
 # document 1.2.0.9000
 
+* check for RStudio in man().
 * Fix return status of write_the_docs().
-* Restructre the Makefile.
 
 # document 1.2.0